requirement.txt
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.12
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==23.1a1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
clip==1.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
coverage==7.0.1
cryptography==39.0.0
cycler==0.11.0
cytoolz==0.12.1
dataclasses==0.6
datasets==2.8.0
-e git+https://github.com/data-centric-ai/dcbench.git@831ab2359d686739d0b0c7a589974ce08448e58d#egg=dcbench
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
docformatter==1.5.1
docutils==0.17.1
domino==0.1.5
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.2
flake8==6.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
huggingface-hub==0.11.1
identify==2.5.11
idna==3.4
imagesize==1.4.1
importlib-metadata==5.2.0
importlib-resources==5.10.2
iniconfig==1.1.1
iopath==0.1.10
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.11.4
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==5.1.1
jupyter-events==0.6.0
jupyter-server==2.0.6
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.12
kaleido==0.2.1
keyring==23.13.1
kiwisolver==1.4.4
lightning-utilities==0.5.0
livereload==2.6.3
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
meerkat-ml==0.2.5
mistune==2.0.4
more-itertools==9.0.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nbsphinx==0.8.10
nest-asyncio==1.5.6
networkx==2.8.8
nodeenv==1.7.0
notebook==6.5.2
notebook-shim==0.2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.9.2
pkgutil-resolve-name==1.3.10
platformdirs==2.6.0
plotly==5.12.0
pluggy==1.0.0
portalocker==2.6.0
pre-commit==2.21.0
progressbar==2.5
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.13.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==7.0.0
pytorch-lightning==1.8.6
pytz==2022.7
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
readme-renderer==37.3
recommonmark==0.7.1
regex==2022.10.31
requests==2.28.1
requests-toolbelt==0.10.1
responses==0.18.0
rfc3986==2.0.0
rich==12.6.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.1.0
tensorboardX==2.5.1
termcolor==2.1.1
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
twine==4.0.2
typing-extensions==4.4.0
ujson==5.6.0
umap-learn==0.5.3
untokenize==0.1.1
urllib3==1.26.13
virtualenv==20.17.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
wget==3.2
widgetsnbextension==4.0.5
xxhash==3.2.0
yacs==0.1.8
yarl==1.8.2
zipp==3.11.0
